requirements.txt
python-telegram-bot[all]
psycopg2
sqlalchemy
loguru
requests
